setting up the thresholds to match with a more realistic backgrounds.

+++ party-photo-booth/src/com/groundupworks/partyphotobooth/image/AndroidImage.java
@@ -13,10 +13,13 @@ public interface AndroidImage {
 	/**
 	 * Check whether the current image is different from the given image.
 	 * A pixel is different if pixel_value - threshold < pixel_threshold < pixel_value + threshold
-	 * An image is different if the total of different pixels is > than threshold.
+	 * An image is different if the total of different pixels is > than threshold
+     * @param background
+     * @param pixel_threshold
+     * @param threshold the percentage (0...1) of the different pixels to consider different image.
 	 */
 	public abstract boolean isDifferent(AndroidImage background, int pixel_threshold, 
-			int threshold);
+			float threshold);
 	
 	/**
 	 * Access the low level data of the image. Data layout is
